+++ Reflection/src/utopia/reflection/component/reach/template/ReachComponentLike.scala
@@ -1,10 +1,13 @@
-package utopia.reflection.component.reach
+package utopia.reflection.component.reach.template
 
-import utopia.genesis.shape.shape2D.Bounds
+import utopia.flow.event.Changing
+import utopia.genesis.shape.shape2D.{Bounds, Point, Size}
 import utopia.genesis.util.Drawer
 import utopia.reflection.component.drawing.template.DrawLevel
-import utopia.reflection.component.drawing.template.DrawLevel._
+import utopia.reflection.component.drawing.template.DrawLevel.{Background, Foreground, Normal}
+import utopia.reflection.component.reach.hierarchy.ComponentHierarchy
 import utopia.reflection.component.template.layout.stack.Stackable2
+import utopia.reflection.text.Font
 
 /**
   * A common trait for "Reach" (no-swing) style components
@@ -16,6 +19,21 @@ trait ReachComponentLike extends Stackable2
 	// ABSTRACT	------------------------
 	
 	/**
+	  * @return A pointer to the current position of this component
+	  */
+	def positionPointer: Changing[Point]
+	
+	/**
+	  * @return A pointer to the current size of this component
+	  */
+	def sizePointer: Changing[Size]
+	
+	/**
+	  * @return A pointer to the current bounds (position + size) of this component
+	  */
+	def boundsPointer: Changing[Bounds]
+	
+	/**
 	  * @return Hierarchy containing all this component's parents
 	  */
 	protected def parentHierarchy: ComponentHierarchy
@@ -25,11 +43,11 @@ trait ReachComponentLike extends Stackable2
 	/**
 	  * Paints the contents of this component on some level. This method is called three times in normal component
 	  * painting process (once for each draw level)
-	  * @param drawer Drawer used for painting this component's area. (0, 0) coordinates should lie at the parent
-	  *               component's top-left corner.
+	  * @param drawer    Drawer used for painting this component's area. (0, 0) coordinates should lie at the parent
+	  *                  component's top-left corner.
 	  * @param drawLevel Targeted draw level (background is drawn first, then normal,
 	  *                  foreground is drawn above child components)
-	  * @param clipZone Limited drawing area. The drawing should be clipped / limited to that area, if specified.
+	  * @param clipZone  Limited drawing area. The drawing should be clipped / limited to that area, if specified.
 	  */
 	protected def paintContent(drawer: Drawer, drawLevel: DrawLevel, clipZone: Option[Bounds] = None): Unit
 	
@@ -45,7 +63,7 @@ trait ReachComponentLike extends Stackable2
 	  * @return The absolute (on-screen) position of this component. None if not connected to main component
 	  *         hierarchy
 	  */
-	def absolutePosition = parentHierarchy.absolutePositionModifier.map { position + _ }
+	def absolutePosition = position + parentHierarchy.absolutePositionModifier
 	
 	/**
 	  * @return The position of this component inside the so called top component
@@ -53,43 +71,61 @@ trait ReachComponentLike extends Stackable2
 	def positionInTop = position + parentHierarchy.positionToTopModifier
 	
 	
+	// IMPLEMENTED	---------------------
+	
+	override def bounds = boundsPointer.value
+	
+	override def position = positionPointer.value
+	
+	override def size = sizePointer.value
+	
+	override def fontMetrics(font: Font) = parentHierarchy.fontMetrics(font)
+	
+	
 	// OTHER	-------------------------
 	
 	/**
 	  * Indicates that this component's and its hierarchy's layout should be updated
 	  */
-	def revalidate() =
-	{
+	def revalidate() = {
 		// Resets the cached stack size of this and upper components
 		resetCachedSize()
 		parentHierarchy.revalidate()
 	}
 	
 	/**
+	  * Paints this component again
+	  */
+	def repaint() = parentHierarchy.repaint(bounds)
+	
+	/**
+	  * Paints this component's parent again
+	  */
+	def repaintParent() = parentHierarchy.repaintBottom()
+	
+	/**
 	  * Paints this component and its children
-	  * @param drawer Drawer to use for drawing this component. Origin coordinates (0,0) should be located at the
-	  *               top-left corner of the parent component.
+	  * @param drawer   Drawer to use for drawing this component. Origin coordinates (0,0) should be located at the
+	  *                 top-left corner of the parent component.
 	  * @param clipZone Bounds where the drawing is limited. None if whole component area should be drawn.
 	  */
-	def paintWith(drawer: Drawer, clipZone: Option[Bounds] = None): Unit =
-	{
+	def paintWith(drawer: Drawer, clipZone: Option[Bounds] = None): Unit = {
 		// Paints background and normal levels
 		paintContent(drawer, Background, clipZone)
 		paintContent(drawer, Normal, clipZone)
 		// Paints child components (only those that overlap with the clipping bounds)
 		val components = children
-		if (components.nonEmpty)
-		{
+		if (components.nonEmpty) {
 			// Calculates new clipping zone and drawer origin
-			val newClipZone = clipZone.map { _ - position }
-			val remainingComponents = newClipZone match
-			{
-				case Some(zone) => components.filter { _.bounds.overlapsWith(zone) }
+			val newClipZone = clipZone.map {_ - position}
+			val remainingComponents = newClipZone match {
+				case Some(zone) => components.filter {_.bounds.overlapsWith(zone)}
 				case None => components
 			}
 			if (remainingComponents.nonEmpty)
 				drawer.translated(position).disposeAfter { d =>
-					remainingComponents.foreach { _.paintWith(d, newClipZone) } }
+					remainingComponents.foreach {_.paintWith(d, newClipZone)}
+				}
 		}
 		// Paints foreground
 		paintContent(drawer, Foreground, clipZone)